requirements.txt
torch>=1.10.2
torchvision>=0.11.3
flwr[simulation]>=0.17
hydra-core>=1.1.1
numpy~=1.22.3
Pillow~=9.0.1